Refactor: Moved ZoneEditor to components folder and updated imports
