+++ src/pages/ZoneEditingPage.jsx
@@ -1,11 +1,11 @@
 import React, { useEffect, useRef, useState } from "react";
 import { useNavigate, useParams } from "react-router-dom";
 import { Stage, Layer, Rect } from "react-konva";
-import { FloorMapService } from "../../services/floormapService.js";
-import "../_pages.scss";
-import ZoneStage from "../../components/ZoneStage/ZoneStage.jsx";
-import ZoneMenu from "../../components/ZoneMenu/ZoneMenu.jsx";
-import { ZoneService } from "../../services/zoneService.js";
+import { FloorMapService } from "../services/floormapService.js";
+import "./_pages.scss";
+import ZoneStage from "../components/ZoneStage/ZoneStage.jsx";
+import ZoneMenu from "../components/ZoneMenu/ZoneMenu.jsx";
+import { ZoneService } from "../services/zoneService.js";
 
 function ZoneEditing() {
   const { floormapId } = useParams();
